FEAT: Updated layout, implemented picture submission and added jQuery
Enhanced several features in the web application. The main updates include the modification of the index and implemented the feature for picture submission in the submit-pictures blade. Moreover, jQuery has been added for improved user interactivity. Other changes include the addition of a new route for submitted reports, restructuring the CSS and the minor tweaks in various modules for optimization.