Add gallery feature with reporting feature
This commit adds a new gallery feature that shows images corresponding to events, dedicating a section to each event-related image. It also enables the use of a context menu for user interaction that provides options to report inappropriate images, copy the image ID, and copy the attachment URL. Moreover, it includes the display of author details in case the image is reported. Consequently, this enhances the user interaction and moderation processes.